Move Navigation component from common to src directory
